# codr-api/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Users demo
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api
...
schemes:
  - http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
paths:
  /right_groups:
    x-swagger-router-controller: right_groups
    get:
      description: Returns a list of all right_groups.
      operationId: getRightGroups
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RightGroup"
        "500":
          description: Internal server error
    post:
      description: Creates a new right_groups.
      operationId: postRightGroup
      parameters:
        - name: body
          in: body
          description: Description of the new right's group.
          required: true
          schema:
            type: object
            properties:
              description:
                type: string
      responses:
        "201":
          description: Success
          schema:
            $ref: "#/definitions/RightGroup"
        "401":
          description: Not allowed
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
  /right_groups/{right_group_id}:
    x-swagger-router-controller: right_groups
    put:
      description: Updates the specified rights group.
      operationId: putRightGroupById
      parameters:
        - name: right_group_id
          in: path
          description: ID of rights group
          required: true
          type: integer
          format: int64
        - name: body
          in: body
          description: Description of the new right's group.
          required: true
          schema:
            type: object
            properties:
              description:
                type: string
      responses:
        "200":
          description: Right group updated
        "401":
          description: Not allowed
        "404":
          description: Id not found
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
    delete:
      description: Deletes the specified rights group.
      operationId: deleteRightGroupById
      parameters:
        - name: right_group_id
          in: path
          description: ID of the right group to delete
          required: true
          type: integer
          format: int64
      responses:
        "204":
          description: Item deleted
        "401":
          description: Not allowed
        "404":
          description: Id not found
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
...
  /modules:
    x-swagger-router-controller: modules
    get:
      description: Returns a list of all modules.
      operationId: getModules
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ModuleList"
        "500":
          description: Internal server error
    post:
      description: Creates a new module.
      operationId: postModule
      parameters:
        - in: body
          name: body
          description: JSON representation of the module to create.
          required: true
          schema:
            type: object
            properties:
              name:
                type: string
              description:
                type: string
      responses:
        "201":
          description: Success, new module has been created.
          schema:
            type: object
            properties:
              id:
                type: integer
                format: int64
              name:
                type: string
              description:
                type: string
        "401":
          description: Error, user is not allowed to access this page.
        "405":
          description: Error, invalid input.
        "500":
          description: Internal server error
  /modules/{module_id}:
    x-swagger-router-controller: modules
    get:
      description: Returns the details of the specified module.
      operationId: getModuleById
      parameters:
        - name: module_id
          in: path
          description: ID of the module to fetch.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ModuleResponse"
        "404":
          description: Error, module could not be found.
          
    delete:
      description: Deletes the specified module.
      operationId: deleteModuleById
      parameters:
        - name: module_id
          in: path
          description: ID of the module to delete.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success.
        "401":
          description: Not allowed.
        "404":
          description: Error, module could not be found.
        "500":
          description: Internal server error
    put:
      description: Updates the specified module.
      operationId: putModuleById
      parameters:
        - in: body
          name: body
          description: Data for the module to update.
          schema:
            $ref: "#/definitions/ModuleUpdate"
        - name: module_id
          in: path
          description: ID of the module to update.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ModuleResponse"
        "401":
          description: Error, user is not allowed to access this page.
        "404":
          description: Error, module could not be found.
        "405":
          description: Incorrect input
        "500":
          description: Internal server error
  /modules/{module_id}/group:
    x-swagger-router-controller: modules
    get:
      description: Returns the groups for this module for the current user.
      operationId: getCurrentUserGroups
      parameters:
        - name: module_id
          in: path
          description: ID of the module from which to fetch groups.
          required: true
          type: integer
          format: int64
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              type: object
              properties:
                group_id:
                  type: integer
                  format: int64
                users:
                  type: array
                  items:
                    $ref: '#/definitions/User'
        "500":
          description: Internal server error
...
definitions:
  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      display_name:
        type: string
...
  Group:
    type: object
    properties:
      id:
        type: integer
        format: int64
      module_id:
        type: integer
        format: int64
      Users:
        type: array
        items: 
          $ref: '#/definitions/User'
...
  ModuleUpdate:
    type: object
    properties:
      name:
        type: string
      description:
        type: string
      users:
        type: array
        items:
          type: object
          properties:
            user_id:
              type: integer
              format: int64
            group_id:
              type: integer
              format: int64
  Module:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      can_view:
        type: boolean
      can_edit:
        type: boolean
  ModuleResponse:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      can_edit:
        type: boolean
      courses:
        type: array
        items:
          $ref: '#/definitions/ModuleCourse'
      groups:
        type: array
        items:
          $ref: '#/definitions/Group'
  ModuleList:
      type: object
      properties:
        can_edit:
          type: boolean
        modules:
          type: array
          items:
            $ref: '#/definitions/Module'
...
  ModuleCourse:
    type: object
    required:
      - name
      - description
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string  
...
  RightGroup:
    type: object
    required:
      - id
      - description
    properties:
      id:
        type: integer
        format: int64
      description:
        type: string
...